+++ data/duke.txt
@@ -1,2 +1,3 @@
 D | 0 | abc | 2004-10-19 1313
 E | 0 | fuck | 2018-10-19 1010 | 2039-11-11 1111
+T | 0 | sfse
