+++ src/main/java/command/AddCommand.java
@@ -1,3 +1,12 @@
+package command;
+
+import core.TaskList;
+import ui.Ui;
+import storage.Storage;
+import exception.BaimiException;
+import task.Task;
+import task.Todo;
+
 public class AddCommand extends Command {
     private String description;
