+++ src/main/java/command/AddEventCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import core.TaskList;
+import ui.Ui;
+import storage.Storage;
+import exception.BaimiException;
+import task.*;
+
 public class AddEventCommand extends Command {
     private String description;
     private String from;
