+++ src/main/java/command/Command.java
@@ -1,3 +1,8 @@
+package command;
+import core.TaskList;
+import ui.Ui;
+import storage.Storage;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws Exception;
     public boolean isExit() {
