+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,12 @@
+package command;
+
+import core.TaskList;
+import ui.Ui;
+import storage.Storage;
+import exception.BaimiException;
+import task.Task;
+import task.Todo;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
