+++ src/main/java/core/Baimi.java
@@ -1,9 +1,9 @@
-import java.io.*;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Scanner;
+package core;
+
+import command.Command;
+import exception.BaimiException;
+import storage.Storage;
+import ui.Ui;
 
 public class Baimi {
     private Storage storage;
