+++ src/main/java/exception/BaimiException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class BaimiException extends Exception {
+    public BaimiException(String message) {
+        super(message);
+    }
+}
