+++ src/main/java/exception/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class EmptyDescriptionException extends BaimiException {
+    public EmptyDescriptionException(String commandType) {
+        super("The description of a " + commandType + " cannot be empty. Please provide a valid description.");
+    }
+}
