+++ src/main/java/exception/InvalidFormatException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class InvalidFormatException extends BaimiException {
+    public InvalidFormatException(String correctFormat) {
+        super("Invalid format. Correct format: " + correctFormat);
+    }
+}
