+++ src/main/java/exception/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class UnknownCommandException extends BaimiException {
+    public UnknownCommandException() {
+        super("Unknown command. Please enter a valid command (e.g., todo, deadline, event, list, delete).");
+    }
+}
