+++ src/main/java/task/Todo.java
@@ -1,3 +1,5 @@
+package task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
